Assume BPB is required and set segment registers
We now changed ORG to 0, and set DS and ES registers to 0x7c0 instead.
We also setup the stack to start at 0x7c00.

We now assume a BPB block needs to be located at the start of the
bootloader, and add it in a way that jumps to our code.