Fix page table retrieval from a pde.
The address in the pde doesn't point to a `PageTable` but rather to the
underlying `PageTableEntry` array. There is no way to find the original
`PageTable` in this context, so we wrap the `PageTableEntry` array with
a `PageTable` by adding a ctor overload that acts as a view and doesn't
clean up the memory (as this will be done by the original `PageTable`).