Refactor interrupts and IDT code
Use generated assembly wrappers for ISRs. The alternative is to use
the `__attribute__((interrupt))` which should work but I don't fully
trust it. It also has the downside that every function you call from
within it has to have `__attribute__((no_caller_saved_registers))` in
the declaration. This requires you to create wrappers for every utility
you use which can end up as more boilerplate.