Refactor paging
Also move allocator to outside of the kernel.
The allocator has a pointer to the underlying implementation which was
created in the `make` function of the kernel. This meant that when the
function returned the `kernel` struct was copied, but the pointer from
the allocator to the implementation still pointed to the copied from
memory causing a corruption when calling malloc/free.
By initializing the allocator implementation in the main we can make
sure that it doesn't get moved and the pointer stays valid. In order to
maintain the abstraction and not pass the specific implementation type
to the kernel's `make` function we also initialize the allocator
interface in the main.